create game_engine folder and move related files there
