+++ xrobot/CMakeLists.txt
@@ -26,7 +26,7 @@ find_package(Boost REQUIRED COMPONENTS python system)
 
 #include_directories(${Boost_INCLUDE_DIRS})
 #include_directories(${PYTHON_INCLUDE_DIRS})
-#link_directories(${Boost_LIBRARY_DIR})
+link_directories(${Boost_LIBRARY_DIR})
 
 ## user can directly set GLOG_INCLUDE_DIRS and GLOG_LIBRARIES
 if(NOT GLOG_LIBRARIES)
@@ -44,103 +44,68 @@ include(glm)
 include(glfw)
 include(bullet)
 
-link_directories(${Boost_LIBRARY_DIR})
-
 set(DEP_LIBS
-    ${GFLAGS_LIBRARIES}
-    ${GLOG_LIBRARIES}
-    ${Boost_LIBRARIES}
-    ${PYTHON_LIBRARIES})
+  ${GFLAGS_LIBRARIES}
+  ${GLOG_LIBRARIES}
+  ${Boost_LIBRARIES}
+  ${PYTHON_LIBRARIES})
 
 set(DEP_INCLUDES
-    ${CMAKE_CURRENT_SOURCE_DIR}
-    ${GFLAGS_INCLUDE_DIRS}
-    ${GLOG_INCLUDE_DIRS}
-    ${Boost_INCLUDE_DIRS}
-    ${PYTHON_INCLUDE_DIRS} )
+  ${CMAKE_CURRENT_SOURCE_DIR}
+  ${GFLAGS_INCLUDE_DIRS}
+  ${GLOG_INCLUDE_DIRS}
+  ${Boost_INCLUDE_DIRS}
+  ${PYTHON_INCLUDE_DIRS})
 
 add_library(utils OBJECT
-    vendor/json.cpp
-    utils.cpp
-)
+  vendor/json.cpp
+  utils.cpp)
+
+if (USE_DOUBLE_PRECISION)
+  add_definitions(-DUSE_DOUBLE_PRECISION)
+endif()
 
 add_subdirectory(render_engine)
 
 add_subdirectory(bullet_engine)
 
-if (USE_DOUBLE_PRECISION)
-  add_definitions(-DUSE_EGL)
-  add_definitions(-DUSE_DOUBLE_PRECISION)
-endif()
+add_subdirectory(game_engine)
 
-set(XROBOT_SOURCES
-    map_grid.cpp
-    map_suncg.cpp
-    state_machine.cpp
-    room_generator.cpp
-    lidar.cpp
-    navigation.cpp
-    inventory.cpp
-    world.cpp
-    python_binding.cpp
-)
+add_library(xrobot SHARED
+  ${XROBOT_SOURCES}
+  $<TARGET_OBJECTS:render_engine>
+  $<TARGET_OBJECTS:bullet_engine>
+  $<TARGET_OBJECTS:game_engine>
+  $<TARGET_OBJECTS:utils>)
 
-set(XROBOT_CPP_EXAMPLE_SOURCES
-    cpp_example/navigation_in_maze.cpp
-    cpp_example/navigation_in_suncg.cpp
-)
-
-if(BUILD_CPP_EXAMPLES)
-    add_library(xrobot SHARED
-        ${XROBOT_CPP_EXAMPLE_SOURCES}
-        ${XROBOT_SOURCES}
-        $<TARGET_OBJECTS:render_engine>
-        $<TARGET_OBJECTS:bullet_engine>
-        $<TARGET_OBJECTS:utils>
-    )
-else()
-    add_library(xrobot SHARED
-        ${XROBOT_SOURCES}
-        $<TARGET_OBJECTS:render_engine>
-        $<TARGET_OBJECTS:bullet_engine>
-        $<TARGET_OBJECTS:utils>
-    )
-endif()
-
-add_dependencies(xrobot utils render_engine bullet bullet_engine)
+add_dependencies(xrobot utils render_engine bullet_engine game_engine)
 
 target_link_libraries(xrobot PUBLIC ${DEP_LIBS})
-
 target_include_directories(xrobot PUBLIC ${DEP_INCLUDES})
 
 set_target_properties(xrobot
   PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR})
 
 if(BUILD_CPP_EXAMPLES)
-    # Navigation in Maze
-    add_executable(xrobot_test_nav cpp_example/example_navigation_in_maze.cpp)
+  # Navigation in Maze
+  add_executable(xrobot_test_nav
+    cpp_example/example_navigation_in_maze.cpp
+    cpp_example/navigation_in_maze.cpp)
+  add_dependencies(xrobot_test_nav xrobot)
+  target_link_libraries(xrobot_test_nav PUBLIC xrobot ${DEP_LIBS})
+  target_include_directories(xrobot_test_nav PUBLIC ${DEP_INCLUDES})
+  set_target_properties(xrobot_test_nav
+    PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/cpp_example)
 
-    add_dependencies(xrobot_test_nav xrobot)
-
-    target_link_libraries(xrobot_test_nav PUBLIC xrobot ${DEP_LIBS})
-
-    target_include_directories(xrobot_test_nav PUBLIC ${DEP_INCLUDES})
-
-    set_target_properties(xrobot_test_nav
-        PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/cpp_example)
-
-    # Navigation in SUNCG
-    add_executable(xrobot_test_suncg cpp_example/example_navigation_in_suncg.cpp)
-
-    add_dependencies(xrobot_test_nav xrobot)
-
-    target_link_libraries(xrobot_test_suncg PUBLIC xrobot ${DEP_LIBS})
-
-    target_include_directories(xrobot_test_suncg PUBLIC ${DEP_INCLUDES})
-
-    set_target_properties(xrobot_test_suncg
-        PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/cpp_example)
-
+  # Navigation in SUNCG
+  add_executable(xrobot_test_suncg
+    cpp_example/example_navigation_in_suncg.cpp
+    cpp_example/navigation_in_suncg.cpp)
+  add_dependencies(xrobot_test_nav xrobot)
+  target_link_libraries(xrobot_test_suncg PUBLIC xrobot ${DEP_LIBS})
+  target_include_directories(xrobot_test_suncg PUBLIC ${DEP_INCLUDES})
+  set_target_properties(xrobot_test_suncg
+    PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/cpp_example)
 endif()
