+++ xrobot/bullet_engine/CMakeLists.txt
@@ -1,20 +1,22 @@
 add_library(bullet_engine OBJECT 
-    common.h
-    common.cpp
-    bullet_joint.cpp
-    bullet_object.cpp
-    bullet_body.cpp
-    bullet_world.cpp
-)
+  common.cpp
+  bullet_joint.cpp
+  bullet_object.cpp
+  bullet_body.cpp
+  bullet_world.cpp)
 
-add_dependencies(bullet_engine assimp bullet glfw glm utils)
+add_dependencies(bullet_engine bullet glm utils)
 
 target_include_directories(bullet_engine PUBLIC
-    ${ASSIMP_INCLUDE_DIRS}
-    ${BULLET_INCLUDE_DIRS}
-    ${GLFW_INCLUDE_DIRS}
-    ${GLM_INCLUDE_DIRS})
+  ${BULLET_INCLUDE_DIRS}
+  ${GLM_INCLUDE_DIRS})
+
+set(DEP_LIBS
+  ${DEP_LIBS}
+  ${BULLET_LIBRARIES}
+  PARENT_SCOPE)
 
 set(DEP_INCLUDES
-    ${DEP_INCLUDES}
-    PARENT_SCOPE)
+  ${DEP_INCLUDES}
+  ${BULLET_INCLUDE_DIRS}
+  PARENT_SCOPE)
