+++ xrobot/game_engine/inventory.cpp
