+++ xrobot/game_engine/inventory.h
