+++ xrobot/game_engine/navigation.cpp
