+++ xrobot/game_engine/world.cpp
